make variable name changes more modular
this change will allow us to string together changes to variable names like this no matter what the data frame we pass in is. this is helpful because it removes a lot of lines of code and gives us more flexbility in name changes like this going forward. it also means this is a lot easier to debug.